Make invite tracker work in the cache too

+++ discord/ext/helpers/invite_tracker.py
@@ -1,20 +1,20 @@
 import asyncio
-from typing import Dict, List, Union
+from typing import Dict, List, Optional, Union
 import aiosqlite
 import discord
 
 
 class InviteTracker:
-    def __init__(self, bot: discord.Client, filename: str) -> None:
+    def __init__(self, bot: discord.Client, database: Optional[str] = None) -> None:
         self.bot = bot
-        self.filename = filename
+        self.database = database
         asyncio.get_event_loop().create_task(self._init())
 
     async def _init(self) -> None:
         self.data = {
             g.id: {i.id: i.uses for i in await g.invites()} for g in self.bot.guilds
         }
-        async with aiosqlite.connect(self.filename) as con:
+        async with aiosqlite.connect(self.database) as con:
             async with con.cursor() as cur:
                 await cur.execute(
                     "CREATE TABLE IF NOT EXISTS invites (guild_id BIGINT, invite_id TEXT, uses INT, PRIMARY KEY (guild_id, invite_id))"
@@ -28,32 +28,34 @@ class InviteTracker:
     async def _add_to_db(
         self, data: Dict[int, Dict[int, int]], init: bool = False, *, guild_id: int = 0
     ) -> None:
-        async with aiosqlite.connect(self.filename) as con:
-            async with con.cursor() as cur:
-                if init:
-                    for guild_id, value in data.items():
-                        for invite_id in value.keys():
+        if self.database is not None:
+            async with aiosqlite.connect(self.database) as con:
+                async with con.cursor() as cur:
+                    if init:
+                        for guild_id, value in data.items():
+                            for invite_id in value.keys():
+                                await cur.execute(
+                                    "INSERT INTO invites VALUES (?, ?, ?)",
+                                    (guild_id, invite_id, data[guild_id][invite_id]),
+                                )
+                    else:
+                        for invite_id, value_ in data.items():
                             await cur.execute(
                                 "INSERT INTO invites VALUES (?, ?, ?)",
-                                (guild_id, invite_id, data[guild_id][invite_id]),
+                                (guild_id, invite_id, value_),
                             )
-                else:
-                    for invite_id, value_ in data.items():
-                        await cur.execute(
-                            "INSERT INTO invites VALUES (?, ?, ?)",
-                            (guild_id, invite_id, value_),
-                        )
 
-                await con.commit()
+                    await con.commit()
 
     async def _remove_from_db(self, what, data) -> None:
-        async with aiosqlite.connect(self.filename) as con:
-            async with con.cursor() as cur:
-                await cur.execute(
-                    f"DELETE FROM invites WHERE {'invite_id' if what == 'invite' else 'guild_id'} = ?",
-                    (data,),
-                )
-                await con.commit()
+        if self.database is not None:
+            async with aiosqlite.connect(self.database) as con:
+                async with con.cursor() as cur:
+                    await cur.execute(
+                        f"DELETE FROM invites WHERE {'invite_id' if what == 'invite' else 'guild_id'} = ?",
+                        (data,),
+                    )
+                    await con.commit()
 
     async def add_guild(self, guild: discord.Guild) -> None:
         self.data[guild.id] = {i.id: i.uses for i in await guild.invites()}
@@ -73,7 +75,7 @@ class InviteTracker:
 
     async def increment_uses(self, invite: discord.Invite, count: int) -> None:
         self.data[invite.guild.id][invite.id] += count
-        async with aiosqlite.connect(self.filename) as con:
+        async with aiosqlite.connect(self.database) as con:
             async with con.cursor() as cur:
                 await cur.execute(
                     "UPDATE invites SET uses = ? WHERE invite_id = ?",
